Restructured dashboard/controller/journal
